requirements.txt
python-telegram-bot==20.0
selenium==4.10.0
requests==2.26.0
fake_useragent==0.1.11
